chore(pivot): fix circular refenrence warning
